requirements.txt.txt
absl-py==0.13.0
AccessControl==5.7
Acquisition==4.13
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
AuthEncoding==4.3
beautifulsoup4==4.9.3
BTrees==4.11.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
Chameleon==3.10.2
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
colorama==0.4.1
cycler==0.10.0
DateTime==4.8
decorator==5.0.9
docopt==0.6.2
DocumentTemplate==4.1
et-xmlfile==1.1.0
ExtensionClass==4.9
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.21.0
idna==2.8
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata==4.5.0
joblib==1.0.1
Js2Py==0.71
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy_loader==0.1
libcst==0.3.19
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.4.2
msgpack==1.0.4
MultiMapping==4.1
multipart==0.2.4
mypy-extensions==0.4.3
numba==0.53.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.0
PasteDeploy==3.0.1
pathspec==0.8.1
Persistence==3.6
persistent==4.9.3
pick==0.6.4
Pillow==8.3.1
pipwin==0.5.1
pooch==1.4.0
prettytable==0.7.2
protobuf==3.20.1rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDrive==1.3.1
pydub==0.25.1
pyfiglet==0.8.post1
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.1
python-gettext==4.1
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.21.0
requests-oauthlib==1.3.0
resampy==0.2.2
RestrictedPython==6.0
roman==3.3
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.15.0
sklearn==0.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
soupsieve==2.2.1
soxr==0.3.3
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.61.2
transaction==3.0.1
typing-extensions==3.7.4.3
typing-inspect==0.7.1
tzdata==2022.7
tzlocal==2.1
unimport==0.8.4
uritemplate==3.0.1
urllib3==1.24.3
waitress==2.1.2
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==2.0.1
windows-curses==2.2.0
wrapt==1.12.1
WSGIProxy2==0.5.1
z3c.pt==3.3.1
zc.lockfile==2.0
ZConfig==3.6.1
zExceptions==4.3
zipp==3.5.0
ZODB==5.8.0
zodbpickle==2.6
